refactor: move utilities to a new file
