+++ example/src/App.tsx
@@ -1,31 +1,15 @@
 import * as React from 'react';
 import { useState } from 'react';
 
-import {
-  StyleSheet,
-  SafeAreaView,
-  Platform,
-  StatusBar,
-  TouchableOpacity,
-  Text,
-  View,
-} from 'react-native';
+import { SafeAreaView, TouchableOpacity, Text, View } from 'react-native';
 import { DraggableList, ScaleDecorator } from 'react-native-x-draggable-list';
 import type { RenderItemParams } from 'react-native-draggable-flatlist';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
+import { COLORS, getBackgroundColor, getCurrentTime, styles } from './utils';
 
-const COLORS = {
-  LIST: ['#FAF1E4', '#CEDEBD', '#9EB384'],
-  DARK: '#435334',
-  SUBTITLE: '#49573c',
-  ACTIVE: '#d6dbd0',
-  LIGHT: '#FAF1E4',
-};
-
-const getBackgroundColor = (index: number) => {
-  return COLORS.LIST[index % COLORS.LIST.length] ?? 'white';
-};
-
+/**
+ * Type of the item in the list
+ */
 export type Item = {
   id: string;
   content: {
@@ -35,6 +19,9 @@ export type Item = {
   };
 };
 
+/**
+ * Prepare example items for the list
+ */
 const getItems = (count: number): Item[] =>
   Array.from({ length: count }, (_, k) => k).map((k) => ({
     id: `item-${k}`,
@@ -44,10 +31,6 @@ const getItems = (count: number): Item[] =>
       color: getBackgroundColor(k),
     },
   }));
-
-const getCurrentTime = () => {
-  return new Date().toLocaleTimeString();
-};
 
 export default function App() {
   const [items, setItems] = useState(getItems(15));
@@ -100,48 +83,3 @@ export default function App() {
     </GestureHandlerRootView>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  backgroundDark: {
-    backgroundColor: COLORS.DARK,
-  },
-  androidSafeArea: {
-    paddingTop: Platform.OS === 'android' ? StatusBar.currentHeight : 0,
-  },
-  rowItem: {
-    height: 100,
-    width: 'auto',
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: 'red',
-  },
-  text: {
-    color: COLORS.DARK,
-    fontSize: 24,
-    fontWeight: 'bold',
-    textAlign: 'center',
-  },
-  subtitle: {
-    color: COLORS.SUBTITLE,
-  },
-  header: {
-    paddingTop: 16,
-    paddingBottom: 16,
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    textAlign: 'center',
-    backgroundColor: COLORS.DARK,
-    color: COLORS.LIGHT,
-    paddingBottom: 10,
-  },
-  logger: {
-    fontSize: 16,
-    color: COLORS.LIGHT,
-    textAlign: 'center',
-  },
-});
